+++ ext/tensor_ext.h
@@ -27,5 +27,6 @@
 #include "tensor/reductions/ref.zep.h"
 #include "tensor/reductions/rref.zep.h"
 #include "tensor/settings.zep.h"
+#include "tensor/tensorbuffer.zep.h"
 
 #endif